moved helpers from routing to lib
