+++ web/src/lib/bot/helpers.ts
